requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
dj-rest-auth==5.0.2
Django==5.0.6
django-allauth==64.1.0
django-filter==24.3
djangorestframework==3.15.2
filelock==3.14.0
idna==3.8
Markdown==3.7
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-environ==0.4.54
requests==2.32.3
setuptools==70.0.0
sqlparse==0.5.0
urllib3==2.2.2
virtualenv==20.26.2
wheel==0.43.0
